+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### History for developement
 
+- 2018/08/20
+    - add FileServer(static) in signal/http.go
+    
 - 2019/08/19
     - read [PLI (Picture Loss Indication)](https://webrtcglossary.com/pli/) over RTCP
     - add Makefile, HISTORY.md, go.{mod.sum}